Tween Duration スケール関数を外部ファイルに切り出した
